maintain: move routing files to a dedicated directory inth app directory
