+++ src/app/routing/app-preloading-strategy.ts
